requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-groq
duckduckgo-search
arxiv
wikipedia
